Move followersDetails & followersList to Followers
